pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ecgbench"
dynamic = ["version"]
description = "Reproducible ECG benchmark datasets with standardised splits, validation, and Croissant metadata"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Vajira Thambawita"}]
keywords = ["ecg", "benchmark", "dataset", "electrocardiogram", "croissant", "mlcommons"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

# Base: lightweight, no heavy deps. Enough for config, catalogue, validation logic, splitting.
dependencies = [
    "numpy>=1.22.0",
    "pandas>=1.4.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.1.0",
]

[project.optional-dependencies]
torch = ["torch>=1.12.0", "wfdb>=4.0.0"]
croissant = ["mlcroissant>=1.0.0"]
download = ["tqdm>=4.60.0"]
hf = ["huggingface_hub>=0.20.0", "python-dotenv>=1.0.0"]
all = ["ecgbench[torch,croissant,download,hf]"]
dev = [
    "ecgbench[all]",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/vlbthambawita/ECGBench"
Repository = "https://github.com/vlbthambawita/ECGBench"
Issues = "https://github.com/vlbthambawita/ECGBench/issues"
Documentation = "https://vlbthambawita.github.io/ECGBench/"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "ecgbench/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["ecgbench"]

[tool.hatch.build.targets.sdist]
include = ["ecgbench/**"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
